+++ src/lua/shared/openutils/OpenItemTweaker/OpenItemTweaker.lua
@@ -4,22 +4,27 @@
 -- that can be found in the LICENSE file.
 --
 
-OpenItemTweaker = OpenItemTweaker or {}
+OpenItemTweaker = {
+    tweaker = nil
+}
 
 function OpenItemTweaker:new()
+    if OpenItemTweaker.tweaker then
+        return OpenItemTweaker.tweaker
+    end
+
     local tweaker = {
-        items = {}
+        items = {},
+        started = false,
     }
 
     -- Apply changes properties for items in list.
     function tweaker.Apply()
-        local item;
-
-        for k,v in pairs(tweaker.items) do
-            for t,y in pairs(v) do
-                item = ScriptManager.instance:getItem(k);
+        for k, v in pairs(tweaker.items) do
+            for t, y in pairs(v) do
+                local item = ScriptManager.instance:getItem(k);
                 if item ~= nil then
-                    item:DoParam(t.." = "..y);
+                    item:DoParam(t.." = " .. y);
                 end
             end
         end
@@ -33,6 +38,28 @@ function OpenItemTweaker:new()
 
         tweaker.items[name][key] = value;
     end
+
+    function tweaker.StartTweaker()
+        if tweaker.started then
+            return
+        end
+
+        tweaker.started = true
+        Events.OnGameBoot.Add(tweaker.Apply);
+    end
+
+    function tweaker.StopTweaker()
+        if not tweaker.started then
+            return
+        end
+
+        tweaker.started = false
+        Events.OnGameBoot.Remove(tweaker.Apply)
+    end
+
+    tweaker.StartTweaker()
+
+    OpenItemTweaker.tweaker = tweaker;
 
     return tweaker
 end